pyproject.toml
[project]
name = "Mephisto"
version = "0.1.0"
description = ""
authors = [
    {name = "nullqwertyuiop", email = "null@member.fsf.org"},
]
dependencies = [
    "toml>=0.10.2",
    "pydantic>=2.5.3",
    "uvicorn>=0.24.0.post1",
    "fastapi>=0.104.1",
    "psutil>=5.9.6",
    "graiax-fastapi>=0.4.1",
    "graia-broadcast>=0.23.4",
    "avilla-core>=1.0.0a17",
    "graia-saya>=0.0.19",
    "aiohttp>=3.9.1",
    "avilla-twilight>=0.3.1",
    "avilla-console>=1.0.0a19",
    "kayaku>=0.5.4",
    "avilla-onebot-v11>=1.0.0a19",
    "richuru>=0.1.1",
    "sqlalchemy>=2.0.29",
    "aiosqlite>=0.20.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "AGPL-3.0 License"}

[project.optional-dependencies]
playwright = [
    "graiax-text2img-playwright>=0.5.0",
    "lxml>=5.1.0",
]
mysql = [
    "aiomysql>=0.2.0",
]

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.isort]
profile = "black"

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.5.0",
    "towncrier>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "black>=23.11.0",
]

[tool.towncrier]
directory = "changelog.d"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
underlines = ["", "", ""]
template = "changelog.d/changelog_template.ninja"
title_format = "## [{version}](https://github.com/nullqwertyuiop/Mephisto/tree/{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/nullqwertyuiop/Mephisto/issues/{issue})"

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "chore"
name = "Other tasks"
showcontent = true
